Almost last

+++ utils.c
@@ -17,10 +17,10 @@ int		bcopy_int(void *dest, void *src, int size)
 	return (0);
 }
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	str = (char *)s;
